+++ scripts/build.sh
@@ -2,25 +2,63 @@
 
 set -e
 
-PATH=$(npm bin):"$PATH"
-export PATH
 CLEANSLATE="node_modules/cleanslate/docs/files/cleanslate.css"
 
+isWindowsMinGW() {
+  local is_mingw="False"
+  if [ "$(uname | cut -c 1-5)" = "MINGW" ]; then
+    is_mingw="True"
+  fi
+
+  echo -n "${is_mingw}"
+}
+
+if [ "$(isWindowsMinGW)" = "True" ]; then
+  WIN_PYTHON="py -3"
+  NPM_BIN_DIR="$(cygpath $(npm bin))"
+  PATH=$NPM_BIN_DIR:$PATH
+else
+  PATH="$(npm bin):$PATH"
+fi
+
+export PATH
+
+mkdir -p build
+mkdir -p build/static
 mkdir -p generated/static
 mkdir -p generated/static/clippy
-scripts/excmds_macros.py
+
+if [ "$(isWindowsMinGW)" = "True" ]; then
+  $WIN_PYTHON scripts/excmds_macros.py
+else
+  scripts/excmds_macros.py
+fi
 scripts/newtab.md.sh
 scripts/make_tutorial.sh
 scripts/make_docs.sh &
-nearleyc src/grammars/bracketexpr.ne > src/grammars/.bracketexpr.generated.ts
-native/install.sh local
 
-(webpack --display errors-only && scripts/git_version.sh)&
+nearleyc src/grammars/bracketexpr.ne \
+  > src/grammars/.bracketexpr.generated.ts
+
+if [ "$(isWindowsMinGW)" = "True" ]; then
+  powershell \
+    -NoProfile \
+    -InputFormat None \
+    -ExecutionPolicy Bypass \
+    native/win_install.ps1 -DebugDirBase native
+else
+  native/install.sh local
+fi
+
+(webpack --display errors-only \
+  && scripts/git_version.sh) &
 
 wait
 
+scripts/bodgecss.sh
+
 if [ -e "$CLEANSLATE" ] ; then
-	cp "$CLEANSLATE" build/static/cleanslate.css
+	cp -v "$CLEANSLATE" build/static/css/cleanslate.css
 else
 	echo "Couldn't find cleanslate.css. Try running 'npm install'"
 fi