+++ src/background.ts
@@ -55,30 +55,52 @@ import * as webext from "./lib/webext"
     l: prom => prom.then(console.log).catch(console.error),
 })
 
+// {{{ Clobber CSP
+
 // This should be removed once https://bugzilla.mozilla.org/show_bug.cgi?id=1267027 is fixed
-let cspListener
-if (config.get("csp") == "clobber") {
-    cspListener = browser.webRequest.onHeadersReceived.addListener(
-        request.addurltocsp,
+function addCSPListener() {
+    browser.webRequest.onHeadersReceived.addListener(
+        request.clobberCSP,
         { urls: ["<all_urls>"], types: ["main_frame"] },
         ["blocking", "responseHeaders"],
     )
 }
+
+function removeCSPListener() {
+    browser.webRequest.onHeadersReceived.removeListener(request.clobberCSP)
+}
+
+config.getAsync("csp").then(csp => csp === "clobber" && addCSPListener())
+
 browser.storage.onChanged.addListener((changes, areaname) => {
-    if (config.get("csp") == "clobber") {
-        cspListener = browser.webRequest.onHeadersReceived.addListener(
-            request.addurltocsp,
-            { urls: ["<all_urls>"], types: ["main_frame"] },
-            ["blocking", "responseHeaders"],
-        )
-    } else {
-        // This doesn't work. :(
-        // browser.webRequest.onHeadersReceived.removeListener(cspListener)
+    if ("userconfig" in changes) {
+        if (changes.userconfig.newValue.csp === "clobber") {
+            addCSPListener()
+        } else {
+            removeCSPListener()
+        }
     }
 })
+
+// }}}
 
 // Prevent Tridactyl from being updated while it is running in the hope of fixing #290
 browser.runtime.onUpdateAvailable.addListener(_ => {})
 
-// Try to load an RC file
-excmds.source_quiet()
+browser.runtime.onStartup.addListener(_ => {
+    config.getAsync("autocmds", "TriStart").then(aucmds => {
+        let hosts = Object.keys(aucmds)
+        // If there's only one rule and it's "all", no need to check the hostname
+        if (hosts.length == 1 && hosts[0] == ".*") {
+            Controller.acceptExCmd(aucmds[hosts[0]])
+        } else {
+            native.run("hostname").then(hostname => {
+                for (let host of hosts) {
+                    if (hostname.content.match(host)) {
+                        Controller.acceptExCmd(aucmds[host])
+                    }
+                }
+            })
+        }
+    })
+})