+++ src/commandline_content.ts
@@ -2,6 +2,7 @@
 
 import Logger from "./logging"
 import * as config from "./config"
+import { theme } from "./styling"
 const logger = new Logger("messaging")
 
 /* TODO:
@@ -34,46 +35,48 @@ async function init() {
             )
             cmdline_iframe.setAttribute("id", "cmdline_iframe")
             hide()
-            window.document.documentElement.appendChild(cmdline_iframe)
+            document.documentElement.appendChild(cmdline_iframe)
+            // first theming of page root
+            await theme(window.document.querySelector(":root"))
         } catch (e) {
             logger.error("Couldn't initialise cmdline_iframe!", e)
         }
     }
 }
 
-// Load the iframe immediately if the document is already complete (happens if tridactyl is reloaded)
+// Load the iframe immediately if we can (happens if tridactyl is reloaded or on ImageDocument)
 // Else load lazily to avoid upsetting page JS that hates foreign iframes.
-if (document.readyState === "complete") {
+try {
     init()
-} else {
+} catch (e) {
     // Surrender event loop with setTimeout() to page JS in case it's still doing stuff.
     document.addEventListener("DOMContentLoaded", () => setTimeout(init, 0))
 }
 
 export function show() {
-    if (enabled) {
+    try {
         const height =
             cmdline_iframe.contentWindow.document.body.offsetHeight + "px"
         cmdline_iframe.setAttribute("style", `height: ${height} !important;`)
-    }
+    } catch (e) {}
 }
 
 export function hide() {
-    if (enabled) {
+    try {
         cmdline_iframe.setAttribute("style", "height: 0px !important;")
-    }
+    } catch (e) {}
 }
 
 export function focus() {
-    if (enabled) {
+    try {
         cmdline_iframe.focus()
-    }
+    } catch (e) {}
 }
 
 export function blur() {
-    if (enabled) {
+    try {
         cmdline_iframe.blur()
-    }
+    } catch (e) {}
 }
 
 export function executeWithoutCommandLine(fn) {