+++ src/config.ts
@@ -76,7 +76,7 @@ const DEFAULTS = o({
         "<c-o>": "back",
         "<c-i>": "forward",
         d: "tabclose",
-        D: "composite tabprev | sleep 100 | tabclose #",
+        D: "composite tabprev; sleep 100; tabclose #",
         gx0: "tabclosealltoleft",
         gx$: "tabclosealltoright",
         u: "undo",
@@ -114,6 +114,7 @@ const DEFAULTS = o({
         ";k": "hint -k",
         ";y": "hint -y",
         ";p": "hint -p",
+        ";P": "hint -P",
         ";r": "hint -r",
         ";s": "hint -s",
         ";S": "hint -S",
@@ -124,6 +125,7 @@ const DEFAULTS = o({
         ";v": "hint -W exclaim_quiet mpv",
         "<S-Insert>": "mode ignore",
         "<CA-Esc>": "mode ignore",
+        "<CA-`>": "mode ignore",
         I: "fillcmdline Ignore mode is now toggled by pressing <S-Insert>",
         a: "current_url bmark",
         A: "bmark",
@@ -131,12 +133,17 @@ const DEFAULTS = o({
         zo: "zoom -0.1 true",
         zz: "zoom 1",
         ".": "repeat",
+        gow: "open http://www.bbc.co.uk/news/live/uk-44167290",
+        gnw: "tabopen http://www.bbc.co.uk/news/live/uk-44167290",
         "<SA-ArrowUp><SA-ArrowUp><SA-ArrowDown><SA-ArrowDown><SA-ArrowLeft><SA-ArrowRight><SA-ArrowLeft><SA-ArrowRight>ba":
             "open https://www.youtube.com/watch?v=M3iOROuTuMA",
     }),
     autocmds: o({
         DocStart: o({
-            "addons.mozilla.org": "mode ignore",
+            // "addons.mozilla.org": "mode ignore",
+        }),
+        TriStart: o({
+            ".*": "source_quiet",
         }),
     }),
     exaliases: o({
@@ -170,6 +177,14 @@ const DEFAULTS = o({
         openwith: "hint -W",
         "!": "exclaim",
         "!s": "exclaim_quiet",
+        colourscheme: "set theme",
+        colours: "colourscheme",
+        colorscheme: "colourscheme",
+        colors: "colourscheme",
+        man: "help",
+        "!js": "js",
+        "!jsb": "jsb",
+        current_url: "composite get_current_url | fillcmdline_notrail ",
     }),
     followpagepatterns: o({
         next: "^(next|newer)\\b|»|>>|more",
@@ -242,6 +257,7 @@ const DEFAULTS = o({
         hinting: 2,
         state: 2,
         excmd: 1,
+        styling: 2,
     }),
     noiframeon: [],
 
@@ -254,6 +270,10 @@ const DEFAULTS = o({
     browser: "firefox",
     nativeinstallcmd:
         "curl -fsSl https://raw.githubusercontent.com/cmcaine/tridactyl/master/native/install.sh | bash",
+    win_powershell_nativeinstallcmd:
+        "Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://raw.githubusercontent.com/gsbabil/tridactyl/master/native/win_install.ps1'))",
+    win_cmdexe_nativeinstallcmd:
+        '@"%SystemRoot%\\System32\\WindowsPowerShell\\v1.0\\powershell.exe" -NoProfile -InputFormat None -Command "iex ((New-Object System.Net.WebClient).DownloadString(\'https://raw.githubusercontent.com/gsbabil/tridactyl/master/native/win_install.ps1\'))"',
     profiledir: "auto",
 
     // Container settings